refactor(coordinator): extract discovery subsystem into DiscoveryMixin
coordinator.py was a 2,421-line god-object; ~1,000 of those lines were
one cohesive concern — the discovery lifecycle. Move it verbatim into a
NikobusDiscoveryMixin (new discovery_mixin.py), mixed in as a base class:

    class NikobusDataCoordinator(NikobusDiscoveryMixin, DataUpdateCoordinator)

What moved: PC-Link inventory + per-module register scans
(start_pc_link_inventory / _try_pclink_inventory / start_module_scan),
phase/progress reporting (_handle_discovery_progress /
_update_discovery_state / discovery_progress_percent), the frame-routing
callbacks (_inventory_callback / _discovery_frame_callback are NOT moved
— they stay with _event_callback; only the post-inventory lifecycle
moves), post-discovery reconciliation (_reconcile_post_discovery +
_surface_legacy_undecoded_buttons), CF ingestion + activation, and the
.nkb name import (async_import_nkb_names + its _output_entity_key /
_apply_entity_name helpers). NKB_IMPORT_CATEGORIES moved to const.py
(shared with config_flow).

Because it's a mixin, 'self' is the same coordinator instance at
runtime: zero behaviour change, every caller and the library's
cross-object contract (get_button_channels, discovery_module) untouched.
A TYPE_CHECKING block on the mixin declares the coordinator surface it
relies on so the file is strict-clean on its own.

coordinator.py 2,421 -> 1,409 (-42%); discovery_mixin.py 1,119.
406 tests pass, ruff clean, mypy 99 (was 100 — errors moved with the
code, no regression).
